+++ test/mocks/storage.h
@@ -28,7 +28,7 @@ namespace {
         MOCK_METHOD(void, write_impl, (const char* buffer, size_t n), (override));
         MOCK_METHOD(void, read_impl, (char* buffer, size_t n), (override));
         MOCK_METHOD(void, seek_impl, (size_t n), (override));
-        MOCK_METHOD(size_t, size_impl, (), (override));
+        MOCK_METHOD(uint64_t, size_impl, (), (override));
     };
 
 } // namespace
